requirements.txt
# SpectraAI Requirements - Universal AI Support

# Core AI Libraries
openai>=1.0.0                    # OpenAI API client (for paid models)
transformers>=4.35.0             # Hugging Face transformers (free local AI)
torch>=2.0.0                     # PyTorch for local AI models
httpx>=0.25.0                    # HTTP client for API calls

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0                  # Data validation and settings

# Async Support
aiofiles>=23.0.0                 # Async file operations

# Optional: Enhanced features
numpy>=1.24.0                    # Numerical operations
scikit-learn>=1.3.0              # ML utilities for emotion analysis

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Type checking and linting (optional)
mypy>=1.0.0
black>=23.0.0
flake8>=6.0.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
